WIP: Add event watcher process to cloud_controller_ng
We have simply cloned what the web process is doing. Now we have a long
running process that has acces to the k8s core resoures and is trying to
watch the LRP events. Unfortunately the watcher is getting a 403
Permission error even though we are making sure that the
serviceaccount it is running with has proper rbac permissions